+++ haproxy/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.7
+
+* Removing sftppass, unsetting consul sysrc parameters where needed
+
+---
+
 0.0.6
 
 * Fixing an overwrite of earlier correction