+++ influxdb/CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.11
+
+* Removing sftppass, unsetting consul sysrc parameters where needed
+
+---
+
 0.0.10
 
 * General security updates